Extracted common argument arsing logic to separate file
