+++ input.hpp
@@ -1,0 +1,3 @@
+int parseKsize(char *arg);
+cv::Mat getImg(char *filename);
+void putImg(cv::Mat image, char *filename);
